refactor: extract action template to pixi module
